Move connection failed dialog
